Revert "netfilter: Rename ingress hook include file"
This reverts commit 0c87db171dec7a5a2f32b1ba1add45724bed00de.
